+++ app/src/main/res/layout/hovedside.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context=".RapportFragment">
+    tools:context=".Hovedside">
 
     <ScrollView
         android:layout_width="match_parent"
@@ -75,26 +75,15 @@
                 android:layout_width="296dp"
                 android:layout_height="48dp"
                 android:layout_marginStart="29dp"
-                android:layout_marginTop="16dp"
+                android:layout_marginTop="8dp"
                 android:layout_marginEnd="29dp"
                 android:elevation="4dp"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintHorizontal_bias="0.48"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@+id/salgitems">
+                app:layout_constraintTop_toBottomOf="@+id/constraintLayout3">
 
             </android.support.constraint.ConstraintLayout>
-
-            <ListView
-                android:id="@+id/salgitems"
-                android:layout_width="363dp"
-                android:layout_height="499dp"
-                android:layout_marginStart="8dp"
-                android:layout_marginTop="8dp"
-                android:layout_marginEnd="8dp"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@+id/constraintLayout3" />
 
 
         </android.support.constraint.ConstraintLayout>
