+++ app/src/main/res/layout/rapport.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context=".HovedsideFragment">
+    tools:context=".Rapport">
 
     <ScrollView
         android:layout_width="match_parent"
@@ -41,10 +41,10 @@
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintHorizontal_bias="0.492"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@+id/Salglistview" />
+                app:layout_constraintTop_toBottomOf="@+id/salgitems" />
 
             <ListView
-                android:id="@+id/Salglistview"
+                android:id="@+id/salgitems"
                 android:layout_width="363dp"
                 android:layout_height="622dp"
                 android:layout_marginStart="8dp"
@@ -67,22 +67,5 @@
 
         </android.support.constraint.ConstraintLayout>
     </ScrollView>
-
-    <Button
-        android:id="@+id/button"
-        android:layout_width="92dp"
-        android:layout_height="92dp"
-        android:layout_marginStart="8dp"
-        android:layout_marginEnd="8dp"
-        android:layout_marginBottom="48dp"
-        android:background="@drawable/roundedbutton"
-        android:text="+"
-        android:textAlignment="center"
-        android:textColor="@android:color/white"
-        android:textSize="40sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.973"
-        app:layout_constraintStart_toStartOf="parent" />
 
 </android.support.constraint.ConstraintLayout>
